glassDet/requirement.txt
numpy>=1.21.0,<2.0.0
opencv-python>=4.5.5.64
pillow>=9.0.0
scipy>=1.7.3
tqdm>=4.64.0
scikit-image>=0.19.2
scikit-learn>=1.0.2
matplotlib>=3.5.1
tensorboard>=2.8.0